完成一级评论展示-倒序排列-页面布局-回复后整体页面刷新

+++ src/main/java/com/community/dto/QuestionDTO.java
@@ -17,7 +17,7 @@ public class QuestionDTO {
     private Long gmtCreate;
     private Long gmt_modified;
     private String creator;
-    private Integer userId;
+    private Long userId;
     private Integer commentCount;
     private Integer viewCount;
     private Integer likeCount;
